==> src/90_Predictions/src/05_NB_IA_Explore_Labels.ipynb <==
## Import Annotations

from utils.database import *
from utils.files import *
from tqdm import tqdm
import pandas as pd

## Read CSV File

path_to_annotations = "../data/20h_experiment/5200-df-full-annotated-feb-and-may-julian-version.csv"
annotations = read_csv_as_dict(path_to_annotations)

df_annotations = pd.DataFrame(annotations)
df_annotations.head()

df_annotations.columns

unique_urls_per_topic = df_annotations.groupby('batch')['url'].nunique()
print(unique_urls_per_topic)

**Unique URLs per topic:** 

unique_urls_per_topic = df_annotations.groupby('topic')['url'].nunique()
print(unique_urls_per_topic)

filtered_df = df_annotations[df_annotations['is_direct_topic_full'] == "False"]
unique_urls_per_topic_filtered = filtered_df.groupby('topic')['url'].nunique()
print(unique_urls_per_topic_filtered)

# Assuming df is your DataFrame
# Create an empty dictionary to store sets of URLs for each topic
urls_per_topic = {}

# Extract unique URLs for each topic
for topic in df_annotations['topic'].unique():
    urls_per_topic[topic] = set(df_annotations[df_annotations['topic'] == topic]['url'].unique())

# Calculate intersections between the sets
# Example for three topics A, B, and C
topic_list = list(urls_per_topic.keys())
if len(topic_list) >= 3:
    common_urls_AB = urls_per_topic[topic_list[0]].intersection(urls_per_topic[topic_list[1]])
    common_urls_AC = urls_per_topic[topic_list[0]].intersection(urls_per_topic[topic_list[2]])
    common_urls_BC = urls_per_topic[topic_list[1]].intersection(urls_per_topic[topic_list[2]])
    common_urls_ABC = urls_per_topic[topic_list[0]].intersection(urls_per_topic[topic_list[1]], urls_per_topic[topic_list[2]])

    print(f"Common URLs between {topic_list[0]} and {topic_list[1]}: {len(common_urls_AB)}")
    print(f"Common URLs between {topic_list[0]} and {topic_list[2]}: {len(common_urls_AC)}")
    print(f"Common URLs between {topic_list[1]} and {topic_list[2]}: {len(common_urls_BC)}")
    print(f"Common URLs across all three topics: {len(common_urls_ABC)}")
else:
    print("Not enough topics to calculate intersections.")


df_annotations['url'].nunique()

# Assuming df_annotations is your DataFrame
# Filter the DataFrame where 'good_for_training' is "False"
filtered_df = df_annotations[df_annotations['good_for_training'] == "False"]

# Group the filtered DataFrame by 'topic' and 'is_direct_topic_full', then count unique 'url'
unique_urls_per_topic_full_filtered = filtered_df.groupby(['topic', 'is_direct_topic_full'])['url'].nunique()

print(unique_urls_per_topic_full_filtered)
