==> src/90_Predictions/src/10_NB_FD_Merge_Page_Text_and_Label.ipynb <==
## Merge Labels and Texts

from utils.database import *
from utils.files import *
from tqdm import tqdm
from bson import ObjectId
import pandas as pd 
import numpy as np

## Load URLs

URLs per batch and topic.

df_labels = pd.read_json('../data/raw/pages_with_labels.json', orient='records', lines=True)
df_labels.head()

print("Number of pages: ", len(df_labels))

df_labels.columns

def get_freqs(dataframe, value_column):
    """Prints the frequency distribution of a specified value column, grouped by 'p_id', as a DataFrame."""
    unique_df = dataframe.drop_duplicates(subset=['p_id', value_column])
    value_counts = unique_df[value_column].value_counts().reset_index()
    value_counts.columns = [value_column, 'Frequency']
    return value_counts

get_freqs(df_labels, 'category')

get_freqs(df_labels, 'good_for_training')

get_freqs(df_labels, 'annotation_type')

## Remove Duplicates

# Removing duplicates based on 'view_url', 'batch_id', and 'topic'
print("Number of pages with labels: ", len(df_labels))
df_labels = df_labels.drop_duplicates(subset=['url', 'batch', 'topic'], keep='first')
print("Number of pages with labels after removing duplicates: ", len(df_labels))

## Load Texts

from datasets import load_dataset, Features, Value, Dataset

# File paths to load texts
file_path = f"../data/raw/pages_with_text.json"
save_path = f"../data/tmp/processed_dataset_all_topics_text_only"

# Define the schema 
features_schema = Features({
    '_id': Value('string'),
    'batch_id': Value('int64'),
    'domain': Value('string'),
    'encoding': Value('string'),
    'file_id': Value('string'),
    'lang': Value('string'),
    'text': Value('string'),
    'text_length': Value('int64'),
    'url': Value('string'),
    'word_count': Value('int64'),
    'view_url': Value('string'),
})

# Load the dataset from the JSON Lines file
dataset = load_dataset('json', data_files=file_path, split='train', features=features_schema, keep_in_memory=False)

print(dataset)

# # Processing the dataset in batches and applying transformations
# def process_batch(batch):
#     return batch # TODO: Returning the batch as is to include all attributes from the JSON objects

# processed_dataset = dataset.map(process_batch, batched=True, batch_size=1000)

# Save the processed dataset to the specified folder
dataset.save_to_disk(save_path)
print(f"Dataset saved to {save_path}")

processed_dataset = dataset
processed_dataset

processed_dataset[0]

## Merge Labels and Text

# Convert the processed dataset to a pandas DataFrame
df_texts = processed_dataset.to_pandas()
df_labels.rename(columns={'url': 'view_url', 'batch': 'batch_id'}, inplace=True)
df_labels.head()

# Merge the text and labels DataFrames
df_texts_and_labels = pd.merge(df_texts, df_labels, 
                               left_on=["batch_id", "view_url"], 
                               right_on=["batch_id", "view_url"], 
                               how='right', suffixes=('_text', '_label'))

# Convert the 'good_for_training' column to boolean
df_texts_and_labels["is_topic"] = df_texts_and_labels["is_direct_topic_full"].map({'True': True, 'False': False})
df_texts_and_labels.columns

df_texts_and_labels[["batch_id", "domain", "lang", "text", "view_url", "topic", "is_topic"]].head(100)

len(df_texts_and_labels)

# Specify the column name
column_name = 'text'

# Count the number of NaN values in the specified column
num_nans_text = df_texts_and_labels[column_name].isna().sum()

# Count the number of non-NaN values in the specified column
num_non_nans_text = df_texts_and_labels[column_name].notna().sum()

# Alternatively, you can also calculate non-NaNs by subtracting NaNs from total elements in the column
total_elements_text = df_texts_and_labels[column_name].size
num_non_nans_text_alternative = total_elements_text - num_nans_text

print(f'Number of NaN values in column "{column_name}": {num_nans_text}')
print(f'Number of non-NaN values in column "{column_name}": {num_non_nans_text}')
print(f'Number of non-NaN values in column "{column_name}" (alternative calculation): {num_non_nans_text_alternative}')

## Remove Pages with same content across Batches

print("Number of pages with labels: ", len(df_texts_and_labels))

# Removing duplicates based on 'view_url', 'batch_id', and 'topic'
#df_texts_and_labels = df_texts_and_labels.drop_duplicates(subset=['view_url', 'text', 'topic'], keep='first')
df_texts_and_labels = df_texts_and_labels.drop_duplicates(subset=['view_url', 'topic'], keep='first') # Only keeping one text per page even if it changes

print("Number of pages with labels after removing duplicates: ", len(df_texts_and_labels))

df_texts_and_labels[["batch_id", "domain", "lang", "text", "view_url", "topic", "is_topic"]].head(100)

## Save Intermediate Datatset

# Convert the merged DataFrame back to a Dataset
dataset_texts_and_labels = Dataset.from_pandas(df_texts_and_labels)

save_path = f"../data/tmp/processed_dataset_all_topics"
dataset_texts_and_labels.save_to_disk(save_path)

dataset_texts_and_labels

## Split Dataset by Topic

topics = ["kinder", "energie", "cannabis"]

for topic in topics:
    
    print("Processing topic: ", topic)
    
    # Filter the dataset for the topic
    dataset_topic = dataset_texts_and_labels.filter(lambda example: example['topic'] == topic, num_proc=16)
    dataset_topic.save_to_disk(f"../data/tmp/processed_dataset_{topic}")
    print("Number of examples in dataset_topic: ", len(dataset_topic))
    
    # Filter the dataset for the topic and positive labels
    dataset_topic_positive = dataset_topic.filter(lambda example: example['is_topic'], num_proc=16)
    dataset_topic_positive.save_to_disk(f"../data/tmp/processed_dataset_{topic}_positive")
    print("Number of examples in dataset_topic_positive: ", len(dataset_topic_positive))
    
    # Filter the dataset for the topic and negative labels
    dataset_topic_negative = dataset_topic.filter(lambda example: not example['is_topic'], num_proc=16)
    dataset_topic_negative.save_to_disk(f"../data/tmp/processed_dataset_{topic}_negative")
    print("Number of examples in dataset_topic_negative: ", len(dataset_topic_negative))
