==> src/90_Predictions/src/15_NB_BD_Buff_Training_Data.ipynb <==
# Buff Training Dataset

from bs4 import BeautifulSoup
from tqdm import tqdm
import pandas as pd 
import os

## Import Data

df = pd.read_csv("../data/train_buff/0000-train_buff.csv")
print(df.head())

## Extract Content from HTML Files

path_prefix = "../data/train_buff/"
extracted_texts = []

for index, row in tqdm(df.iterrows(), total=len(df)):
    file_path = path_prefix + row['full_path']
    
    # Ensure the file exists (optional, based on your use case)
    if os.path.exists(file_path):
        
        # Open and read the HTML file
        with open(file_path, 'r', encoding='utf-8') as file:
            html_content = file.read()
        
        # Parse the HTML content with BeautifulSoup
        soup = BeautifulSoup(html_content, 'html.parser') 
        
        text_content = soup.get_text(separator=' ', strip=True)
        extracted_texts.append(text_content)
        #print(text_content)
        
        # Example: Find and print the title tag content
        #print(soup.title.text)
    else:
        print(f"File not found: {file_path}")
        extracted_texts.append(None)

df['extracted_text'] = extracted_texts

df['topic'] = df['topic_folder'].str.replace('_html', '')

df.head()

## Save as Dataset

from datasets import Dataset, Features, ClassLabel, Value, Sequence, DatasetDict

df['topics'] = df['topic'].apply(lambda topic: {topic: {'is_topic_bool': True}})
df.head()

from urllib.parse import urlparse


def extract_domain(url: str) -> str:
    """
    Extracts the domain from a given URL. Prepends 'http://' if no scheme is found to ensure correct parsing. Returns an empty string if the URL is invalid or empty.
    """
    if not url:
        return ""
    
    if not urlparse(url).scheme:
        url = "http://" + url
    
    return urlparse(url).netloc.split('/')[0]

print(extract_domain("www.example.com/path/to/resource"))  # Output: www.example.com
print(extract_domain(""))

df['topics'] = df['topic'].apply(lambda topic: {topic: {'is_topic_bool': True}})
                                            
# Transform the DataFrame to match the desired Hugging Face Dataset format
transformed_data = {
    '_id': df['id'].apply(lambda x: 'dummy_id_' + str(x)),
    'batch_id': 99999,  
    'domain': [extract_domain(url) for url in df['url'].to_list()],  # Extract domain from URL
    'view_url': df['url'], 
    'lang': 'de', 
    'text': df['extracted_text'],
    'text_length': df['extracted_text'].apply(len),
    'word_count': df['extracted_text'].apply(lambda x: len(x.split())),
    'topics': df['topics'].tolist(),
    'topic': df['topic'],  
    'category': "buff",
    'good_for_training': 'True', 
    'good_for_augmentation': 'True', 
    'annotation_type': 'buff',
    'is_topic': True,
    'label': 1, 
}

# Create a new DataFrame that aligns with the Hugging Face Dataset format
hf_df = pd.DataFrame(transformed_data)

# Convert the pandas DataFrame to a Hugging Face dataset
hf_dataset = Dataset.from_pandas(hf_df)

print(hf_dataset)

hf_dataset[0]

hf_dataset.save_to_disk('../data/tmp/processed_dataset_all_topics_buff')

## Save per Topic

topic = "cannabis"
filtered_dataset = hf_dataset.filter(lambda example: example['topic'] == topic)
constant_dict = {topic: {"is_topic_bool": True}}
filtered_dataset = filtered_dataset.map(lambda example: {"topics": constant_dict}, batched=False)
filtered_dataset = filtered_dataset.remove_columns(['label'])
filtered_dataset.save_to_disk(f'../data/tmp/processed_dataset_{topic}_buff')

filtered_dataset[0]

topic = "kinder"
filtered_dataset = hf_dataset.filter(lambda example: example['topic'] == topic)
constant_dict = {topic: {"is_topic_bool": True}}
filtered_dataset = filtered_dataset.map(lambda example: {"topics": constant_dict}, batched=False)
filtered_dataset = filtered_dataset.remove_columns(['label'])
filtered_dataset.save_to_disk(f'../data/tmp/processed_dataset_{topic}_buff')

topic = "energie"
filtered_dataset = hf_dataset.filter(lambda example: example['topic'] == topic)
constant_dict = {topic: {"is_topic_bool": True}}
filtered_dataset = filtered_dataset.map(lambda example: {"topics": constant_dict}, batched=False)
filtered_dataset = filtered_dataset.remove_columns(['label'])
filtered_dataset.save_to_disk(f'../data/tmp/processed_dataset_{topic}_buff')
