==> src/90_Predictions/src/15_NB_BD_Combine_with_Buff.ipynb <==
## Combine Dataset with Training Data Buff

from collections import Counter
from datasets import load_from_disk, Dataset, ClassLabel, Value, Features, load_dataset
from transformers import AutoTokenizer
import numpy as np
import pandas as pd
import random
import torch

import logging
from transformers import logging as transformers_logging

# Set the logging level to error for transformers, which will suppress warnings
transformers_logging.set_verbosity_error()


**Load Examples:**

topic = "cannabis" #"energie" #"kinder" "cannabis"

columns = ['_id', 'batch_id', 'domain', 'view_url', 'lang', 'text', 'text_length', 'word_count', 'topic', 'category', 'good_for_training', 'good_for_augmentation','annotation_type', 'is_topic']

def select_columns(dataset: Dataset, columns_to_keep: list) -> Dataset:
    """Returns a new dataset containing only the specified columns."""
    # Directly compute columns to remove and apply removal
    return dataset.remove_columns([col for col in dataset.column_names if col not in columns_to_keep])


dataset_orig = load_from_disk(f"../data/tmp/processed_dataset_{topic}")
dataset_orig = dataset_orig.select_columns(columns)

print(dataset_orig)
print(dataset_orig[1])

dataset_buff = load_from_disk(f"../data/tmp/processed_dataset_{topic}_buff")
dataset_buff

dataset_buff = load_from_disk(f"../data/tmp/processed_dataset_{topic}_buff")
dataset_buff = dataset_buff.select_columns(columns)
print(dataset_buff)
print(dataset_buff[0])

**Merge Datasets:**

from datasets import concatenate_datasets

dataset_orig[0]

# Define the function to update the records
def update_record(example):
    example['_id'] = str("999999") if example['_id'] is None else example['_id']
    example['domain'] = '' if example['domain'] is None else example['domain']
    example['view_url'] = '' if example['view_url'] is None else example['view_url']
    example['lang'] = '' if example['lang'] is None else example['lang']
    example['text'] = '' if example['text'] is None else example['text']

    # Dynamically calculate text_length and word_count from the text content
    example['text_length'] = int(len(example['text']))
    example['word_count'] = int(len(example['text'].split()))

    example['annotation_type'] = 'not_annotated' if example['annotation_type'] is None else example['annotation_type']
    return example

# Apply the function to the dataset
dataset_orig = dataset_orig.map(update_record)

dataset_orig = dataset_orig.cast_column('text_length', Value('int64'))
dataset_orig = dataset_orig.cast_column('word_count', Value('int64'))


dataset_orig[0]

dataset_buff[0]

dataset = concatenate_datasets([dataset_orig, dataset_buff])

dataset[0]

# Convert boolean labels to integers (True to 1, False to 0)
dataset = dataset.map(lambda example: {'label': int(example['is_topic'])})

# Define a ClassLabel feature for the converted integer labels
class_label_feature = ClassLabel(num_classes=2, names=['False', 'True'])

# Update the features of the dataset
new_features = dataset.features.copy()
new_features['label'] = class_label_feature

# Cast the dataset to the new features
dataset = dataset.cast(new_features)

# Verify the dataset
print(dataset)

# Access the ClassLabel feature for the 'label' column
class_label_feature = dataset.features['label']

# Get the names (string representations)
label_names = class_label_feature.names
print("Label Names:", label_names)

# Create a mapping from integer indices to names
label_mapping = {index: class_label_feature.int2str(index) for index in range(class_label_feature.num_classes)}
print("Label Mapping:", label_mapping)

# To access the name for a specific label index
label_index = 0  # Example index
label_name = class_label_feature.int2str(label_index)
print(f"Label for index {label_index}:", label_name)


from urllib.parse import urlparse

def extract_domain_from_view_url(example):
    
    # Check if 'view_url' is empty
    if not example['view_url']:
        # Set 'view_url' to the value of 'domain'
        example['view_url'] = example['domain']
        
    # Ensure the URL has a protocol for urlparse to work correctly
    view_url = example['view_url']
    if not urlparse(view_url).scheme:
        view_url = "http://" + view_url  # Prepend with a default protocol
    
    # Extract domain using urlparse on the modified URL
    parsed_url = urlparse(view_url)
    
    # Sometimes the domain might be in 'path' if 'netloc' is empty (missing protocol)
    domain = parsed_url.netloc if parsed_url.netloc else parsed_url.path
    
    # Splitting by "/" in case the URL without protocol directly starts with domain name
    domain = domain.split('/')[0]
    
    # Update the 'domain' field
    example['domain'] = domain
    return example

# Apply the function to each example in the dataset
updated_dataset = dataset.map(extract_domain_from_view_url)

dataset[0]

len(dataset)

# Count exaples with label = 1 
label_count = Counter(dataset_orig['is_topic'])
print("Label Count orig:", label_count)

label_count = Counter(dataset_buff['is_topic'])
print("Label Count buff:", label_count)

label_count = Counter(dataset['label'])
print("Label Count combined:", label_count)

## Save to Disk

file_path = f"../data/tmp/processed_dataset_{topic}_buffed"
dataset.save_to_disk(file_path)
#dataset.to_json(file_path)
