==> src/90_Predictions/src/50_NB_TC_Prepare_Dataset.ipynb <==
## Train Multiclass Classifier: BERT

from transformers import AutoTokenizer, AutoModelForSequenceClassification, AutoModelForMaskedLM, TrainingArguments, Trainer
from datasets import Dataset, load_dataset, load_from_disk, concatenate_datasets
from sklearn.metrics import accuracy_score
from collections import Counter
import random
import numpy as np
import torch
import os

topics = ["cannabis", "energie", "kinder"]

## Load Dataset

**Map class-names to class-ids:**

id_to_class = {0: "other",1: "cannabis", 2: "energie", 3: "kinder"}
class_to_id = {"other": 0, "cannabis": 1, "energie": 2, "kinder": 3}

**Load dataset for each topic:**

dataset_cannabis = load_from_disk(f"../data/tmp/processed_dataset_cannabis_buffed_filtered")
dataset_energie = load_from_disk(f"../data/tmp/processed_dataset_energie_buffed_filtered")
dataset_kinder = load_from_disk(f"../data/tmp/processed_dataset_kinder_buffed_filtered")

**Update dataset schema and class-label mappings:**

from datasets import ClassLabel, Features, Value, DatasetDict, concatenate_datasets, load_dataset
import datasets

# Define the new class label feature
class_labels = ClassLabel(names=["other", "cannabis", "energie", "kinder"])

# Define the new features, including all existing ones plus the updated 'label'
new_features = datasets.Features({
    '_id': datasets.Value('string'),
    'batch_id': datasets.Value('int64'),
    'domain': datasets.Value('string'),
    'view_url': datasets.Value('string'),
    'lang': datasets.Value('string'),
    'text': datasets.Value('string'),
    'text_length': datasets.Value('int64'),
    'word_count': datasets.Value('int64'),
    'token_count': datasets.Value('int64'),
    'topic': datasets.Value('string'),
    'category': datasets.Value('string'),
    'good_for_training': datasets.Value('string'),
    'good_for_augmentation': datasets.Value('string'),
    'annotation_type': datasets.Value('string'),
    'is_topic': datasets.Value('int64'),
    'label': class_labels # Updated ClassLabel feature
})


dataset_cannabis = dataset_cannabis.map(lambda e: {'label': class_to_id['cannabis'] if e["label"] == 1 else class_to_id['other']}, features=new_features)

dataset_energie = dataset_energie.map(lambda e: {'label': class_to_id['energie'] if e["label"] == 1 else class_to_id['other']}, features=new_features)

dataset_kinder = dataset_kinder.map(lambda e: {'label': class_to_id['kinder'] if e["label"] == 1 else class_to_id['other']}, features=new_features)

**Merge the three datasets:**

# Concatenate all datasets
dataset_all_topics = concatenate_datasets([dataset_cannabis, dataset_energie, dataset_kinder])

dataset_all_topics

dataset_all_topics[0]

**Filter out annotations not good for training:**

dataset_all_topics_test = dataset_all_topics.filter(lambda example: example['good_for_training'] == "False")

dataset_all_topics = dataset_all_topics.filter(lambda example: example['good_for_training'] == "True")

**Separate negative and positive examples:**

# Filter positive and negative examples
dataset_all_topics_pos = dataset_all_topics.filter(lambda example: example['label'] > 0, num_proc=16)
dataset_all_topics_neg = dataset_all_topics.filter(lambda example: example['label'] == 0, num_proc=16)

**Get examples which are negative for all topics:**

# Collect all view_url values from dataset_all_topics_pos
pos_view_urls = set(dataset_all_topics_pos['view_url'])

# Filter dataset_all_topics_neg to exclude any rows present in dataset_all_topics_pos
dataset_all_topics_neg = dataset_all_topics_neg.filter(lambda example: example['view_url'] not in pos_view_urls, num_proc=16)


dataset_all_topics_neg

**Deduplicate negative examples:**

seen_urls = set()

dataset_all_topics_neg = dataset_all_topics_neg.filter(lambda example: example['view_url'] not in seen_urls and not seen_urls.add(example['view_url']), num_proc=16)

## Oversample Positive Exmaples

MODEL_NAME = "FacebookAI/xlm-roberta-base"
tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
model = AutoModelForMaskedLM.from_pretrained(MODEL_NAME).eval()

if torch.cuda.device_count() > 1:
    print(f"Using {torch.cuda.device_count()} GPUs!")
    model = torch.nn.DataParallel(model)

# Move model to GPU if available
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
model = model.to(device)

def randomly_replace_tokens(text, tokenizer, model, mask_probability=0.15):
    """Elegantly replace tokens one by one, each with full context."""

    inputs = tokenizer(text, return_tensors='pt', truncation=True, padding='max_length', add_special_tokens=True)
    input_ids = inputs.input_ids.clone()
    attention_mask = inputs.attention_mask

    # Identify non-special tokens for potential masking
    non_special_token_indices = [i for i, token_id in enumerate(input_ids[0])
                                 if token_id not in (tokenizer.cls_token_id, tokenizer.sep_token_id, tokenizer.pad_token_id)]
    
    # Randomly select tokens for masking
    num_tokens_to_mask = int(len(non_special_token_indices) * mask_probability)
    tokens_to_mask = np.random.choice(non_special_token_indices, size=num_tokens_to_mask, replace=False)

    for i in tokens_to_mask:
        original_token_id = input_ids[0, i].item()  # Save the original token ID
        masked_input_ids = input_ids.detach().clone()
        masked_input_ids[0, i] = tokenizer.mask_token_id  # Mask the token

        with torch.no_grad():
            outputs = model(masked_input_ids, attention_mask=attention_mask)

        predictions = outputs.logits[0, i]
        predictions[original_token_id] = -float('Inf')  # Invalidate the original token
        best_pred_idx = predictions.argmax(dim=-1).item()
        input_ids[0, i] = best_pred_idx  # Replace with the best prediction

    replaced_text = tokenizer.decode(input_ids[0], skip_special_tokens=True)
    return replaced_text

# Example usage
text = "Das hier ist ein Test."
replaced_text = randomly_replace_tokens(text, tokenizer, model, mask_probability=0.35)
print("Original text:", text)
print("Replaced text:", replaced_text)

from datasets import concatenate_datasets, Dataset

def oversample(dataset, label_column, tokenizer, model, mask_probability=0.35):
    """Oversample minority classes in a Hugging Face Dataset of text data.
    """
    # Calculate class distributions and find the maximum count
    class_counts = Counter(dataset['label'])
    max_class_count = round(max(class_counts.values()) * 1.5) #TODO: Change back
    
    # Initialize a list to hold original and synthetic samples
    oversampled_datasets = []
    
    # Process each class
    for label, count in class_counts.items():
        
        class_indices = [i for i, x in enumerate(dataset[label_column]) if x == label]
        class_dataset = dataset.select(class_indices)
        
        # Calculate how many new samples are needed
        num_new_samples = max_class_count - count
        
        if num_new_samples > 0:
            
            # Randomly select samples to be duplicated and modified
            indices_to_augment = np.random.choice(class_indices, size=num_new_samples, replace=True)
            samples_to_augment = dataset.select(indices_to_augment)
            
            # Apply the randomly_replace_tokens function to generate new samples
            augmented_dataset = samples_to_augment.map(
                lambda example: {"text": randomly_replace_tokens(example['text'], tokenizer, model, mask_probability=mask_probability)},
                batched=False
            )
            
            # Append augmented dataset to the list
            oversampled_datasets.append(augmented_dataset)
        
        # Always include the original dataset for this class
        oversampled_datasets.append(class_dataset)
    
    # Concatenate all datasets into a new Dataset
    return concatenate_datasets(oversampled_datasets)


# Count the occurrences of each label
label_counts = Counter(dataset_all_topics_pos['label'])
print("Class frequencies before:", label_counts)

dataset_all_topics_pos = oversample(dataset_all_topics_pos, 'label', tokenizer, model)

# Count the occurrences of each label
label_counts = Counter(dataset_all_topics_pos['label'])
print("Class frequencies after:", label_counts)

## Sample Negative Examples

print("Number of distinct negative domains", len(set(dataset_all_topics_neg["domain"])))

# Count the occurrences of each label
label_counts = Counter(dataset_all_topics_pos['label'])
print("Class frequencies:", label_counts)

# Find the minimum count
min_count = min(label_counts.values())
print("Minimum class frequency:", min_count)

import random
from datasets import Dataset

def sample_random_from_dataset(dataset, n=5):
    """Samples n random examples from the dataset and returns both the sampled and unsampled parts.
    """

    # Select the sampled dataset
    n = min(n, len(dataset))
    random_indices = random.sample(range(len(dataset)), n)
    sampled_dataset = dataset.select(random_indices)
    
    # Determine indices not included in the random sample to create the remainder dataset
    all_indices = set(range(len(dataset)))
    unsampled_indices = list(all_indices - set(random_indices))
    remaining_dataset = dataset.select(unsampled_indices)
    
    return sampled_dataset, remaining_dataset

print(len(dataset_all_topics_pos))

dataset_all_topics_neg, dataset_all_topics_neg_test = sample_random_from_dataset(dataset_all_topics_neg, n = len(dataset_all_topics_pos))

dataset_all_topics = concatenate_datasets([dataset_all_topics_pos, dataset_all_topics_neg])
dataset_all_topics_holdout = concatenate_datasets([dataset_all_topics_test, dataset_all_topics_neg_test])

seen_urls = set()

dataset_all_topics_holdout = dataset_all_topics_holdout.filter(lambda example: example['view_url'] not in seen_urls and not seen_urls.add(example['view_url']), num_proc=16)

## Sample Example for Validation Set

datasets = dataset_all_topics.train_test_split(test_size=0.05, shuffle=True)
datasets["valid"] = datasets["test"]
datasets["test"] = dataset_all_topics_holdout

datasets.save_to_disk(f"../data/tmp/processed_dataset_all_topics_multiclass")

## Chunkify Examples

from langchain.text_splitter import RecursiveCharacterTextSplitter, CharacterTextSplitter

MODEL_NAME = "FacebookAI/xlm-roberta-base"
MAX_CONTENT_LENGTH = 384
OVERLAP = 64

# Load a pre-trained tokenizer
tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)

def get_input_length(text):
    """ Tokenize the input text and return the number of tokens """
    return len(tokenizer.encode(text, add_special_tokens=True, truncation=False, padding=False))

print(get_input_length("Hello, my name is John Doe"))

text_splitter = RecursiveCharacterTextSplitter(
    chunk_size=MAX_CONTENT_LENGTH,
    chunk_overlap=OVERLAP,
    length_function=get_input_length,
    separators = ["\n\n", "\n", ".", "?", "!", " ", ""]
)

# text_splitter = CharacterTextSplitter(
#     separator = ".", # Split text by sentences
#     chunk_size=MAX_CONTENT_LENGTH,
#     chunk_overlap=OVERLAP,
#     length_function=get_input_length,
#     is_separator_regex=False,
# )

test_text = datasets['train'][1]['text']
print(test_text[:100])

texts = text_splitter.split_text(test_text)
print("Number of Chunks:", len(texts)) 
print("First Chunk:",texts[0])
print("Length of text:", get_input_length(texts[0]))


def expandRow(row, col_name):
    """
    Generate prompts based on text chunks from the input row.
    """
    rows = []

    # Split the text into chunks
    text_chunks = text_splitter.split_text(row.get(col_name, ""))

    # Generate prompts for each text chunk
    for chunk_id, text_chunk in enumerate(text_chunks):
        new_row = {
            **row, 'chunk_id': chunk_id, 'text': text_chunk
        }
        rows.append(new_row)

    return rows

from multiprocessing import Pool


def processDataset(dataset, num_processes, func, params=()):
    """Process a list of articles in parallel using a multiprocessing Pool."""

    # Creates a list of arguments for each call to func
    # Uses starmap to pass multiple arguments to func
    with Pool(processes=num_processes) as pool:
        args = [(row,) + params for row in dataset]
        dataset = list(pool.starmap(func, args))

    # Flatten the resulting list of lists
    # and convert it into a Dataset
    dataset = [item for sublist in dataset for item in sublist]
    dataset = Dataset.from_dict(
        {key: [dic[key] for dic in dataset] for key in dataset[0]})

    return dataset

datasets

num_processes = 24
params = ("text",)

for split in datasets:
    datasets[split] = processDataset(datasets[split], num_processes, expandRow, params)

label_counts = Counter(datasets["train"]['label'])
print("Class frequencies:", label_counts)

datasets

**Update text length:**

def update_metrics(example):
    example['text_length'] = len(example['text'])
    example['word_count'] = len(example['text'].split())
    example['token_count'] = get_input_length(example['text'])
    return example

datasets = datasets.map(update_metrics)

datasets["train"][0]

**Extract URL path:**

from urllib.parse import urlparse, urlunparse

def extract_url_path(example):
    view_url = example['view_url']
    if "://" not in view_url:
        view_url = "http://" + view_url  # Assume http if no protocol specified
    parsed_url = urlparse(view_url)
    new_url = urlunparse(('', '', parsed_url.path, parsed_url.params, parsed_url.query, parsed_url.fragment))
    example['url_path'] = new_url.lstrip('/')  # Store the result in a new field
    return example


extract_url_path({"view_url": "https://www.google.com/search?q=python+url+path"})

datasets = datasets.map(extract_url_path) 

## Save Dataset to Disk

datasets.save_to_disk(f"../data/tmp/processed_dataset_multiclass_chunkified_{MAX_CONTENT_LENGTH}")
