==> src/90_Predictions/src/52_NB_PD_Apply_Self_Training.ipynb <==
## Train Multiclass Classifier: BERT

from transformers import AutoTokenizer, AutoModelForSequenceClassification, AutoModelForMaskedLM, TrainingArguments, Trainer
from datasets import Dataset, load_dataset, load_from_disk, concatenate_datasets
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
from collections import Counter
import random
import numpy as np
import torch
import os

topics = ["cannabis", "energie", "kinder"]

## Load Dataset

**Map class-names to class-ids:**

id_to_class = {0: "other",1: "cannabis", 2: "energie", 3: "kinder"}
class_to_id = {"other": 0, "cannabis": 1, "energie": 2, "kinder": 3}

MAX_CONTENT_LENGTH = 384
file_path = f"../data/tmp/processed_dataset_multiclass_chunkified_{MAX_CONTENT_LENGTH}"
dataset = load_from_disk(file_path)

dataset = dataset["train"]

dataset

## Load Model

model_path = "../models/bert_multiclass_model_buff_incr_neg"
tokenizer = AutoTokenizer.from_pretrained(model_path)
model = AutoModelForSequenceClassification.from_pretrained(model_path).eval()

if torch.cuda.device_count() > 1:
    print(f"Using {torch.cuda.device_count()} GPUs!")
    model = torch.nn.DataParallel(model)

# Move model to GPU if available
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
model.to(device)

## Prepare Dataset

# Tokenize the text
def tokenize_function(examples):
    return tokenizer(examples["text"], padding="max_length", truncation=True) # TODO: URL REMOVED

dataset = dataset.map(tokenize_function, batched=True)

## Get Predictions

def predict_batch(batch):
    """ Perform prediction on a batch of samples in a multiclass setting. """
    
    # Ensure input tensors are on the correct device
    input_ids = torch.tensor(batch['input_ids']).to(device)
    attention_mask = torch.tensor(batch['attention_mask']).to(device)
    
    # Perform prediction
    with torch.no_grad():
        outputs = model(input_ids=input_ids, attention_mask=attention_mask)
        predictions = torch.nn.functional.softmax(outputs.logits, dim=-1)
    
    # Extract probabilities for all classes and predicted classes
    batch['probas'] = predictions.cpu().tolist()  # Move results back to CPU and convert to list
    batch['preds'] = torch.argmax(predictions, dim=-1).cpu().tolist()

    return batch


dataset = dataset.map(predict_batch, batched=True, batch_size=512)

id = 100
print(dataset[id]["label"])
print(dataset[id]["preds"])
print(dataset[id]["probas"])

#dataset.save_to_disk(file_path + "_preds")

## Compare Before and After Filtering

from collections import Counter
from tqdm import tqdm

**Count Difference in Chunk Level Predictions:**

preds = dataset["preds"]
labels = dataset["label"]

print(Counter(preds))
print(Counter(labels))

**Compare URL Level Predictions (loss of Acc after self-training):**

# Group dataset examples by URL, with a fallback to domain
grouped_dataset = {}
for example in tqdm(dataset):
    url = example.get("view_url") or example.get("domain")
    example_filtered = {k: example[k] for k in ["text", "domain", "probas", "preds", "label", "category", "annotation_type", "lang"]}
    grouped_dataset.setdefault(url, []).append(example_filtered)

# Extract labels
labels_for_url = []
for url, chunks in grouped_dataset.items():
    preds = [chunk["label"] for chunk in chunks]
    labels_for_url.append(max(preds))

def majority_voting(answers):
    """Apply majority voting to a list of arbitrary classification answers."""
    count = Counter(answers)
    most_common = count.most_common()  # Get all common answers sorted by frequency

    if not most_common:
        return 0 # Handle empty input scenario

    # Check for ties at the highest count
    max_votes = most_common[0][1]
    tied_classes = [cls for cls, votes in most_common if votes == max_votes]

    if len(tied_classes) > 1:
        return max(tied_classes)  # Return the maximum class label in case of a tie
    return tied_classes[0]  # Return the class with the most votes

majority_voting([1, 1, 2, 2, 2, 3])

prediction_for_url_max = []

# It is kinda nuts that max() works best but the intuition is that most chunks are other and as soon there is anything higher than other (class label 0) it is always related to that policy
for url, chunks in grouped_dataset.items():
    preds = [chunk["preds"] for chunk in chunks]
    #preds = preds if len(preds) > 0 else [0]
    pred = majority_voting([pred for pred in preds if pred > 0]) if max(preds) > 0 else 0
    prediction_for_url_max.append(pred)

# Assuming labels and preds are lists or arrays containing the true labels and predicted labels respectively
accuracy = accuracy_score(labels_for_url, prediction_for_url_max)
precision_per_class = precision_score(labels_for_url, prediction_for_url_max, average=None)
recall_per_class = recall_score(labels_for_url, prediction_for_url_max, average=None)
f1_per_class = f1_score(labels_for_url, prediction_for_url_max, average=None)

print("Overall Accuracy: {:.2f}%".format(accuracy * 100))
print("Precision per class: {}".format(np.round(precision_per_class, 2)))
print("Recall per class: {}".format(np.round(recall_per_class, 2)))
print("F1 Score per class: {}".format(np.round(f1_per_class, 2)))

**Analyse Label Divergence:**

# Sample data (replace these with your actual datasets)
preds = dataset["preds"]
labels = dataset["label"]

# Data preparation
transition_counts = {}
for label, pred in zip(labels, preds):
    if label not in transition_counts:
        transition_counts[label] = {}
    if pred not in transition_counts[label]:
        transition_counts[label][pred] = 0
    transition_counts[label][pred] += 1


import pandas as pd

data = []
for label, preds in transition_counts.items():
    for pred, count in preds.items():
        if label != pred:
            data.append({'Label': label, 'Prediction': pred, 'Count': count})
            
# Sort by label and prediction for better readability
df = pd.DataFrame(data)
df.sort_values(by=['Label', 'Prediction'], inplace=True)
print(df)

**Examples with deviant Predictions:**

from tabulate import tabulate

urls = list(grouped_dataset.keys())
chunks = list(grouped_dataset.values())
table_data = []  # List to store all row data for tabulation

for id_x, (label, prediction) in enumerate(zip(labels_for_url, prediction_for_url_max)):
    if label !=  0 and prediction == 0:
        # Display only up to n chunks, each with detailed information
        for chunk_id, chunk in enumerate(chunks[id_x][:25]):
            preds = chunk['preds']
            text = chunk['text']
            token_count = len(text.split(" "))

            # Limiting text output to the first 50 and last 50 tokens
            text_preview = (text[:50] + '...' + text[-50:]) if len(text) > 100 else text

            # Append each chunk as a separate row in the table
            table_data.append([
                id_x + 1,  # Example ID
                urls[id_x][:250],
                id_to_class[label],
                id_to_class[prediction],
                chunk_id + 1,
                preds,
                token_count,
                text
            ])

        # Break the loop 
        if len(table_data) >= 10: 
            break

# Formatting the output using tabulate
headers = ['Example ID', 'URL', 'Label', 'Prediction', 'Chunk ID', 'Predictions', 'Token Count', 'Text Preview']
display(tabulate(table_data, headers=headers, tablefmt='html', maxcolwidths=[5, 10, 5,  5, 5, 5, 5, 45]))


## Filter Training Set

filtered_dataset = dataset.filter(lambda example: example['label'] == example['preds'] or example['label'] == 0)

dataset

filtered_dataset

MAX_CONTENT_LENGTH = 384
file_path = f"../data/tmp/processed_dataset_multiclass_chunkified_{MAX_CONTENT_LENGTH}"
dataset = load_from_disk(file_path)

#dataset = dataset["train"]

dataset["train"] = filtered_dataset

dataset.save_to_disk(file_path + "_filtered")
